feature: Move handlers to a package
- Reorganises handlers into a separate package
