+++ go.mod
@@ -6,3 +6,5 @@ require (
 	github.com/go-chi/chi/v5 v5.0.11 // indirect
 	golang.org/x/net v0.20.0 // indirect
 )
+
+replace handlers => /handlers
