+++ handlers/go.mod
@@ -1,0 +1,3 @@
+module skull/server/handlers
+
+go 1.21.6
