+++ handlers/jsonHandling.go
@@ -1,0 +1,29 @@
+package handlers
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+)
+
+type User struct {
+	Firstname string `json:"firstname"`
+	Lastname  string `json:"lastname"`
+	Age       int    `json:"age"`
+}
+
+func HandleJsonDecode(w http.ResponseWriter, r *http.Request) {
+	var user User
+	json.NewDecoder(r.Body).Decode(&user)
+	fmt.Fprintf(w, "%s %s is %d years old!", user.Firstname, user.Lastname, user.Age)
+}
+
+func HandleJsonEncode(w http.ResponseWriter, r *http.Request) {
+	peter := User{
+		Firstname: "John",
+		Lastname:  "Doe",
+		Age:       25,
+	}
+
+	json.NewEncoder(w).Encode(peter)
+}
