+++ handlers/websocket.go
@@ -1,0 +1,11 @@
+package handlers
+
+import (
+	"io"
+
+	"golang.org/x/net/websocket"
+)
+
+func EchoServer(ws *websocket.Conn) {
+	io.Copy(ws, ws)
+}
