+++ main.go
@@ -1,42 +1,19 @@
 package main
 
 import (
-	"encoding/json"
-	"fmt"
-	"html/template"
-	"io"
 	"net/http"
 	"time"
+
+	"skull/server/handlers"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
 	"golang.org/x/net/websocket"
 )
 
-type Todo struct {
-	Title string
-	Done  bool
-}
-
-type TodoPageData struct {
-	PageTitle string
-	Todos     []Todo
-}
-
-type User struct {
-	Firstname string `json:"firstname"`
-	Lastname  string `json:"lastname"`
-	Age       int    `json:"age"`
-}
-
-func EchoServer(ws *websocket.Conn) {
-	io.Copy(ws, ws)
-}
-
 func main() {
 
 	r := chi.NewRouter()
-	tmpl := template.Must(template.ParseFiles("templates/base.html"))
 
 	// A good base middleware stack
 	r.Use(middleware.RequestID)
@@ -55,51 +32,16 @@ func main() {
 	})
 
 	// template endpoint
-	r.Get("/todo/{listName}", func(w http.ResponseWriter, r *http.Request) {
-
-		listName := chi.URLParam(r, "listName")
-
-		// create the page data
-		data := TodoPageData{
-			PageTitle: listName,
-			Todos: []Todo{
-				{Title: "Task 1", Done: false},
-				{Title: "Task 2", Done: false},
-				{Title: "Task 3", Done: true},
-			},
-		}
-
-		// render the data in the template
-		tmpl.Execute(w, data)
-	})
+	r.Get("/todo/{listName}", handlers.HandleTodoList)
 
 	// url parameters endpoint
-	r.HandleFunc("/books/{title}/page/{page}", func(w http.ResponseWriter, r *http.Request) {
-
-		title := chi.URLParam(r, "title")
-		page := chi.URLParam(r, "page")
-
-		fmt.Fprintf(w, "You've requested the book: %s on page %s\n", title, page)
-
-	})
+	r.HandleFunc("/books/{title}/page/{page}", handlers.HandleBookRequest)
 
 	// == endpoints for json encoding ==
-	r.HandleFunc("/json/decode", func(w http.ResponseWriter, r *http.Request) {
-		var user User
-		json.NewDecoder(r.Body).Decode(&user)
-		fmt.Fprintf(w, "%s %s is %d years old!", user.Firstname, user.Lastname, user.Age)
-	})
-	r.HandleFunc("/json/encode", func(w http.ResponseWriter, r *http.Request) {
-		peter := User{
-			Firstname: "John",
-			Lastname:  "Doe",
-			Age:       25,
-		}
+	r.HandleFunc("/json/decode", handlers.HandleJsonDecode)
+	r.HandleFunc("/json/encode", handlers.HandleJsonEncode)
 
-		json.NewEncoder(w).Encode(peter)
-	})
-
-	r.Handle("/ws/echo", websocket.Handler(EchoServer))
+	r.Handle("/ws/echo", websocket.Handler(handlers.EchoServer))
 
 	// set up static file serving
 	fs := http.FileServer(http.Dir("static/"))
